graphql request in resolvers unified

+++ schema/mutation/deleteBookmarkMutation.ts
@@ -1,5 +1,8 @@
 import { GraphQLID } from "graphql";
 import { BookmarkType } from "../types/bookmarkType";
+import { GraphQLError } from "graphql";
+
+import { RequestWithAuth } from "../middleware/isAuth";
 
 const Bookmark = require("../../mongoModels/bookmarkSchema");
 
@@ -8,7 +11,12 @@ export const deleteBookmarkMutationField = {
   args: {
     id: { type: GraphQLID },
   },
-  resolve(_source: unknown, args: { id: string }) {
+  resolve(_source: unknown, args: { id: string }, request: RequestWithAuth) {
+    if (!request.isAuth) {
+      return new GraphQLError("Auth error");
+      // throw new Error("Auth error");
+    }
+
     return Bookmark.findByIdAndDelete(args.id);
   },
 };
